Fix missed switch (missing default) branch cases
The branch analysis depth-ordering was bad, it was putting the nested
branch blocks after the return block.  I think I originally did this to
avoid dealing with "for loops" which cause the CFG to have cycles, but
the logic was pretty complex.

This time I manually broke for loops by modifying the cfg builder
results and rewrote it as a simple "repeat until stable" loop which
seems to have done the trick.